requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==23.2.1
langchain
PyMuPDF==1.24.3
Motor==3.1.1
pymongo==4.3.3
pydantic-mongo==2.3.0
qdrant-client==1.10.1
openai==1.58.1
pydantic[email]
python-jose[cryptography]
bcrypt==4.0.1
passlib==1.7.4
cohere==5.14.0
mangum
gunicorn
packaging
